// frontend/src/lib/api/generated/models/registerProjectInputBody.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { PlatformIdentityPayload } from "./platformIdentityPayload.ts";

export interface RegisterProjectInputBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  default_branch?: string;
  display_name?: string;
  local_path: string;
  platform_identity?: PlatformIdentityPayload;
}