// frontend/src/lib/api/generated/models/registerWorktreeInputBody.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface RegisterWorktreeInputBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  base_dir?: string;
  base_ref?: string;
  branch: string;
  create_on_disk?: boolean;
  path?: string;
  setup_script?: string;
  worktree_name?: string;
}